requirements.txt
pyrogram
TgCrypto
Pyromod
pymongo
dnspython
pytz
tzdata
backports-zoneinfo